Dedup sqlite cf_link codec + sqliteQuery factory (08-open-questions #24)
Three duplications consolidated, with the cell.ts ↔ cf-link.ts cycle broken via
a cycle-free leaf module:

- NEW builtins/sqlite/cf-link-codec.ts — the PURE codec: `encodeCellToSigilString`
  (cell→sigil string) and `parseCfLinkToSigil` (parse/validate prologue). Depends
  only on link-utils + types (no runtime cell.ts import), so cell.ts can import it
  without a cycle.
- `encodeSqliteParams` (cell.ts) and `encodeCfLinkValue` (cf-link.ts) now both
  call `encodeCellToSigilString` instead of inlining `createSigilLinkFromParsedLink`
  + JSON.stringify — so write-path and read-path sigils stay byte-identical by
  construction.
- The bound-cell recovery (`isCell` || `toCell` back-pointer) is now a single
  exported `asBoundCell` in cell.ts (where `isCell`/CellImpl live); cf-link.ts
  imports it (same direction it already imported `isCell`, no new cycle),
  replacing its private `asCellOrUndefined`.
- `decodeCfLinkValue` now delegates to `parseCfLinkToSigil` (one prologue, not
  two copies); cf-link.ts re-exports `parseCfLinkToSigil` so existing importers
  are unchanged.
- NEW builtins/sqlite/query-node.ts — the single `sqliteQuery` node factory,
  imported by both `db.query` (cell.ts) and the `sqliteQuery` builder export
  (built-in.ts), so they construct the same node instead of each calling
  createNodeFactory. (Leaf module → no cell.ts ↔ built-in.ts cycle.)

Net −52 lines. Runner sqlite unit + e2e (11/0), cell-core/as-cell green;
typecheck + lint clean.
